Repackage the script and add a test.
Update the package with a pyroject.py file so it can be easily
installed directly from the repo. Add a test in anticipation of
refactoring the script.

To get the script to work with python3.9.13 I had to switch to using
PdfMerger instead of PdfFileMerger. Otherwise a test run would bomb
with this error:

    PyPDF2.errors.DeprecationError: PdfFileMerger is deprecated and was removed in PyPDF2 3.0.0. Use PdfMerger instead.

I also trimmed down the page size assumption. I'm guessing 6000 always
works with production data but my test files are way smaller. I'm
guessing that any value above 0 would work fine here.

The initial test works around the script as it is today. This should
provide a smooth path to refactor it into something a bit more
flexible and command-line oriented.